% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyEmulators.r
\name{applyEmulators}
\alias{applyEmulators}
\title{Apply the emulators to a dataset (e.g. as produced by a deterministic simulation model)}
\usage{
applyEmulators(Df, itrain, itest, responseVariables = "y",
  predictorVariables = NULL, returnSeFit = FALSE, tuneModels = TRUE,
  tunedParams = NULL, emulatorNames = c("lm", "gam", "svm", "nnet",
  "xgb", "locfit", "gp"), verbose = FALSE)
}
\arguments{
\item{Df}{A \code{data.frame} object to apply the emulators to}

\item{itrain}{Indices of the training set (i.e. row numbers of \code{Df}).}

\item{itest}{Indices of the test set (i.e. row numbers of \code{Df}).}

\item{responseVariables}{A string, or vector or strings, giving the name(s) of the response variable(s) to be used in the emulators. These should appear as variable names in the \code{Df} argument. If one string is provided, all emulators will use the same response variable; if this argument is a vector of strings, it should have the same length as the \code{emulatorNames} argument (below). The default value is \code{'y'}.}

\item{predictorVariables}{A vector of strings giving the names of the predictor variables to be used by the emulators. These should appear as variable names in the \code{Df} argument. The default value is \code{NULL}, which results in all variable names from \code{Df} other than those specified by \code{responseVariables}.}

\item{returnSeFit}{Calculate and return the `native' estimate of uncertainty from the emulators, where available (i.e., the standard error of the fit, the width of the confidence interval, or the posterior standard deviation of the estimate). This adds considerably to the time and memory requirements for the Gaussian process regression, in particular. The default value is \code{FALSE}.}

\item{tuneModels}{Boolean: should hyperparameter estimation be undertaken? The default value is \code{TRUE}. If \code{TRUE}, then the resulting hyperparameters are returned.}

\item{tunedParams}{If the hyperparameter estimation has already been undertaken for this dataset, this can be provided as a list. The default value is \code{NULL}, in which case models are}

\item{emulatorNames}{A vector of strings giving the names of the emulators, defaults to \code{c('lm','gam','svm','nnet','xgb','locfit', 'gp').}}

\item{verbose}{Boolean: should extra output be printed to the R console? Defaults to \code{FALSE}.}
}
\value{
A \code{list} object with the following 2 or 3 entries (there will be 3 entries if argument \code{returnSeFit} is \code{TRUE}, and 2 otherwise). \code{pred}: a \code{data.frame} object with column names given by \code{emulatorNames}, and the rows corresponding to \code{itest}. \code{tunedParams}: the list of tuned hyperparameters. \code{predSe}: the native estimate of uncertainty from each emulator for which it is available (this return value is only included if \code{returnSeFit} is \code{TRUE}); this has the same format as the return value of \code{pred} (a \code{data.frame} object with column names given by \code{emulatorNames}, and the rows corresponding to \code{itest}).
}
\description{
Apply the emulators to a dataset (e.g. as produced by a deterministic simulation model)
}
