% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fracText.r
\name{fracText}
\alias{fracText}
\title{Place text on a plot, with locations given as fraction from a corner rather than with absolute coordinates}
\usage{
fracText(fx, fy, labels, ...)
}
\arguments{
\item{fx}{A fraction (or a vector of fractions) of the plot interval across from the left to be. These should be numeric and range from 0.0 to 1.0.}

\item{fy}{A fraction (or a vector of fractions) of the plot interval across from the bottom to be. These should be numeric and range from 0.0 to 1.0. It should have the same length as \code{fx}.}

\item{labels}{A vector of strings. If \code{fx} and \code{fy} have length greater than 1, then \code{labels} should have the same length.}

\item{...}{Further arguments passed to \code{text()}.}
}
\value{
Nothing is returned
}
\description{
Place text on a plot, with locations given as fraction from a corner rather than with absolute coordinates
}
