% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTests.r
\name{runTests}
\alias{runTests}
\title{Run tests and produce plots required for this analysis}
\usage{
runTests(Model = NULL, Ix = NULL, outFile = NULL,
  plotsFolder = getwd())
}
\arguments{
\item{Model}{A string or vector of strings, determining which emulator model to run. The entries for this argument should be some combination of of \code{c('lorenz63', 'lorenz96', 'ODEperiodic')}. The defualt value is \code{NULL}, all three models are run.}

\item{Ix}{The index of the model to be the focus of the emulator uncertainty analysis. For \code{Model = 'lorenz63'}, options for \code{Ix} are 1, 2, 3; for \code{Model = 'lorenz96'}, options are 1, 4, 8, 12, and for \code{Model = 'ODEperiodic'}, the options for \code{Ix} are 1, 2, 3, 4. If \code{NULL}, then all the options are run for the simulation model specified by \code{Model}.}

\item{outFile}{A string specifying the file path to output the standard error results to. If \code{NULL} (the default), the progress will be printed on the R console.}

\item{plotsFolder}{A string specifying the folder to save the plots file to. If unspecified, it defaults to the current working directory.}
}
\value{
Nothing is returned
}
\description{
Run tests and produce plots required for this analysis
}
